(maint) Fetch any new tags from the remote in pa_matchbatch.sh
If you point to an already cloned repo that hasn't fetched from the
remote in a while, there's a chance that it may not have any new tags
that have been created. This causes ticketmatch to output JIRA tickets
from the releases that pushed those tags as tickets that have not been
marked with the right fix version.