(maint) Handle moved tickets in ticketmatch.rb
When a ticket has moved in JIRA, but a commit already references that
ticket identifier, the commit was being flagged as unmatched. This is
fixed in this commit by running an extra JIRA query of the old
identifier and mapping the (real) key that gets returned. The script
output will specify that the ticket identifier is an old alias.