(maint) Improve JIRA team filter
This commit reworks how the (optional) Jira team parameter is used. Previously,
we would only search for tickets in the given project where 'Team' was set to
the given value. This resulted in a lot of entries for `Git commits in Jira`,
since other teams' commits were found in git but not in the Jira query (since
they were filtered out). Now we search all tickets, but only report on the given
team's tickets when identifying unresolved tickets and tickets without release
notes.