(maint) Remove "Release Notes" field
In Puppet's previous Jira instance, tickets had a "Release Notes" field
that indicated what type of change associated with the ticket (i.e.
whether the change was a "Security Fix," "Enhancement," etc.).

(The "Release Notes" field is distinct from the "Release Notes Summary"
field, which contains a description of the change.)

In Perforce's Jira, there is no field equivalent to the "Release Notes"
field from Puppet's previous Jira, so we remove all references to it
from the ticketmatch script.