(maint) Don't examine Epics for release notes
Perforce's Jira instance does not display the "Release Note Summary"
field to users on Epics the way a typical ticket does.

Because of this, ticketmatch.rb flags Epics as not having release notes,
even though a user cannot add any through the web interface as they
would expect.

This commit adds logic to determine if a ticket is an Epic, then ignores
when Epics do not have release notes.