(maint) Fix version parsing in private repos
Previously, Ticketmatch would look for a component's version file using
the name of the directory as the component name. When attempting to
parse the version file of a private repository, this caused Ticketmatch
to fail to find the version file and fall through to using the Git log
to determine the component's version.

This commit updates Ticketmatch to strip the "-private" from a
component's name before attempting to parse the version file.